move the PDP p2s classes to their own file
